Recreated plugin with official angular-library project structure
