--- .github/workflows/main.yml
@@ -1,32 +1,0 @@
-name: Build library
-
-on: [push]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v1
-      ## cache dependencies
-      - name: Cache dependencies
-        uses: actions/cache@v1
-        with:
-          path: node_modules
-          key: ${{ runner.OS }}-build-${{ hashFiles('**/package-lock.json') }}
-          restore-keys: |
-            ${{ runner.OS }}-build-${{ env.cache-name }}-
-            ${{ runner.OS }}-build-
-            ${{ runner.OS }}-
-      ## set up node.js
-      - name: Set up node.js
-        uses: actions/setup-node@v1
-        with:
-          node-version: 13.1
-      ## install dependencies
-      - name: Install dependencies
-        run: npm install
-      ## build package
-      - name: Run build
-        run: npm run build
