--- ng-package.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "./dist",
-  "lib": {
-    "entryFile": "src/index.ts",
-    "umdModuleIds": {
-      "rete": "rete"
-    }
-  }
-}
