+++ projects/angular-render-plugin/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "rete-angular-render-plugin",
+  "version": "0.1.1",
+  "peerDependencies": {
+    "@angular/common": "7 - 8",
+    "@angular/core": "7 - 8",
+    "@angular/elements": "7 - 8",
+    "rete": "^1.4.2"
+  }
+}
