+++ projects/angular-render-plugin/src/lib/components/node/node.component.html
