+++ projects/angular-render-plugin/src/lib/components/node/node.component.sass
@@ -1,4 +1,4 @@
-@import "../vars"
+@import "../../theme/vars"
 
 .node
   background: $node-color
