+++ projects/angular-render-plugin/src/lib/components/socket/socket.component.html
@@ -1,0 +1,1 @@
+<div *ngIf="socket" [ngClass]="[type, socket.name]" [title]="socket.name" class="socket"></div>
