+++ projects/angular-render-plugin/src/lib/components/socket/socket.component.sass
@@ -1,4 +1,4 @@
-@import "../vars"
+@import "../../theme/vars"
 
 \:host
   display: inline-block
