+++ projects/angular-render-plugin/src/lib/components/socket/socket.component.ts
@@ -1,10 +1,10 @@
 import {Component, Input} from '@angular/core';
 import {Input as ReteInput, IO, Socket} from 'rete';
-import {SocketType} from '../types';
+import {SocketType} from '../../types';
 
 @Component({
   selector: 'rete-socket',
-  template: `<div *ngIf="socket" class="socket" [ngClass]="[type, socket.name]" [title]="socket.name"></div>`,
+  templateUrl: './socket.component.html',
   styleUrls: ['./socket.component.sass']
 })
 export class SocketComponent {
