+++ projects/angular-render-plugin/src/lib/pipes/kebab.pipe.ts
