+++ projects/angular-render-plugin/src/public-api.ts
@@ -1,0 +1,6 @@
+import {NodeComponent} from './lib/components/node/node.component';
+
+export {AngularRenderPlugin} from './lib/plugin';
+export {ReteModule} from './lib/module';
+export * from './lib/types';
+export {NodeComponent} from './lib/components/node/node.component';
